requirements.txt
astroid==2.4.2
attrs==20.2.0
dj-database-url==0.5.0
Django==2.2.9
django-debug-toolbar==3.1.1
gunicorn==20.0.4
iniconfig==1.0.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.5.0
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.0.2
pytest-django==3.8.0
pytz==2020.1
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.2.0
wrapt==1.12.1
